+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - more controller tests
 
+### Fixed
+- modification methods in controller
+
 ## 2019-10-11
 ### Fixed
 - method to modify book